#1100 WIP moving eg2 to dynamo0.3 domain
