+++ examples/lfric/eg2/module_inline_example.py
@@ -1,7 +1,7 @@
 # -----------------------------------------------------------------------------
 # BSD 3-Clause License
 #
-# Copyright (c) 2020, Science and Technology Facilities Council.
+# Copyright (c) 2017-2021, Science and Technology Facilities Council
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,9 +30,43 @@
 # LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
 # ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 # POSSIBILITY OF SUCH DAMAGE.
-# ------------------------------------------------------------------------------
-# Author: A. R. Porter, STFC Daresbury Laboratory
+# -----------------------------------------------------------------------------
+# Author R. Ford STFC Daresbury Lab
+# Modified by S. Siso, and A. R. Porter, STFC Daresbury Laboratory
 
-EXAMPLES=$(sort $(wildcard eg*) inline)
+''' Example showing the use of the module-inline transformation for the
+    LFRic domain. '''
+from __future__ import print_function
+import os
+from psyclone.parse.algorithm import parse
+from psyclone.psyGen import PSyFactory, Kern
+from psyclone.transformations import KernelModuleInlineTrans
 
-include ../top_level.mk
+
+def inline():
+    ''' Function exercising the module-inline transformation. '''
+    _, info = parse(os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                                 "single_invoke.x90"),
+                    api="dynamo0.3")
+    psy = PSyFactory("dynamo0.3").create(info)
+    invokes = psy.invokes
+    print(psy.invokes.names)
+    invoke = invokes.get("invoke_0_testkern_type")
+    schedule = invoke.schedule
+    schedule.view()
+    kern = schedule.walk(Kern)[0]
+    # setting module inline directly
+    kern.module_inline = True
+    schedule.view()
+    # unsetting module inline via a transformation
+    trans = KernelModuleInlineTrans()
+    trans.apply(kern, {"inline": False})
+    schedule.view()
+    # setting module inline via a transformation
+    trans.apply(kern)
+    schedule.view()
+    print(str(psy.gen))
+
+
+if __name__ == "__main__":
+    inline()
